coap_client: make existing golioth_get_block_cb_fn internally defined
The `golioth_get_block_cb_fn` callback is defined in client.h but only used
in coap_client.c, which is an internal file. This commit renames the
typedef to `coap_get_block_cb_fn` and moves it to coap_client.h.

We will use `golioth_get_block_cb_fn` for public blockwise download
functions, but it needs a different set of members so it will be
reimplemented in the next commit.

Signed-off-by: Mike Szczys <mike@golioth.io>