blockwise: include transfer context struct by value
The post_block and get_block contexts had been including a
reference to a transfer context struct that is allocated
separately. We can remove an allocation and simplify error
handling by including the full struct as a member rather
than including a pointer.

This also fixes a structurally dead code report from
Coverity (that was also actually a memory leak).

Signed-off-by: Sam Friedman <sam@golioth.io>