zephyr: client: use ATOMIC_DEFINE for defining the  variable
Zephyr's atomic functions expect an array of `atomic_t`, but
we were passing a pointer to a single `atomic_t`, which worked
in practice because the number of flags was less than the size
of `atomic_t`, but this triggered an issue in Coverity for
potential out-of-bounds memory access.

Instead, use Zephyr's `ATOMIC_DEFINE()` macro to define the
`client_flags` variable, as recommended.

Signed-off-by: Sam Friedman <sam@golioth.io>