zephyr: coap: reset block context on first block2 response
Reset the block context on the first BLOCK2 response instead of
on the last BLOCK1 request.

Zephyr's search for options in a CoAP packet stops either when it
encounters a payload marker or when it reaches the end of the
maximum length of the CoAP packet buffer. On receiving response
packets, Zephyr sets the maximum length of the packet buffer to the
actual length of the packet. However, when constructing request
packets, the actual length of the CoAP packet is frequently smaller
than the maximum length of the buffer. This means that searching
for options in a request packet without a payload (such as a GET)
will often search past the end of the actual packet and attempt to
parse garbage data.

When we added support for blockwise responses to blockwise POSTs,
we needed to reset the block context after the POST finishes so
that we can reuse the structure for the response. We did this by
searching for a BLOCK1 option with more_data==false in the request,
but as outlined above, if the BLOCK1 option is not present, the
search continues past the end of the CoAP packet, leading to error
messages about CoAP options exceeding the maximum length. This
search was bounded by the maximum length of the request packet
buffer, so there was no impact to functionality, but the error
logs rightly caused concerns for customers.

Now, instead of searching for the end of a blockwise POST (i.e.
a BLOCK1 option with more_data==false), we search for the
beginning of a blockwise response (i.e. a BLOCK2 option with
block_num==0) to know when we need to reset the block context.
This is functionally equivalent, and because received CoAP
responses have their maximum length set to the actual size of
the CoAP packet, the option search will not proceed past the end
of the packet.

Signed-off-by: Sam Friedman <sam@golioth.io>