port: zephyr: update openthread to resolve deprecation warnings
openthread is removing the instance member of openthread_context which is
causing deprecation warnings at build time. Replace references to this
context member with a pointer to the instance.

Zephyr currently lacks the "get" function needed to acquire the single
instance pointer. But it has been added in the openthread project. This
commit cannot be merged until that change has been merged into the Zephyr
version of the library.

https://github.com/openthread/openthread/commit/e83f3e11b1e13077b122c288d8f49262e1129828

resolves https://github.com/golioth/firmware-issue-tracker/issues/857

Signed-off-by: Mike Szczys <mike@golioth.io>