zephyr: fw_update: add rpi_pico2 as a build-only target
Note that this uses the board name rpi/pico2/rp2350a/m33/w/mcuboot (with
the additional mcuboot qualifier at the end).

Twister will pass locally for the fw_update sample, but fails to find a
valid image when run in CI. This board has been added as build-only until
the issue is resolved.

Signed-off-by: Mike Szczys <mike@golioth.io>