hil: lightdb: remove delete step to stabilize test
we delete a LightDB state endpoint immediately before setting up the set
values. This was shown to have caused failures as the first path we set
after the delete operation was frequently not actually set on the server.
Because we are creating new devices for each test, there is no need to run
the delete. It has been removed to stabilize the test.

Signed-off-by: Mike Szczys <mike@golioth.io>