zephyr: avoid using %.*s in logging
The Zephyr logging system copies data to use later for printing the log
message. When doing so, the string pointer length is determined using
`strlen()`, ignoring the supplied length. If the memory pointed to is not
null-terminated, this results in reading past the end of the array. Avoid
using `%.*s` with Zephyr.

See:
- https://github.com/zephyrproject-rtos/zephyr/issues/93999
- https://github.com/zephyrproject-rtos/zephyr/issues/57867

Signed-off-by: Mike Szczys <mike@golioth.io>