zcbor_utils: gracefully end partially decoded map
If a map contains entries which are not decoded after the last
entry which is decoded, then the map will not be ended gracefully.
This can interfere with further processing of elements in the
CBOR message. This commit ensures that we process the map until
we reach the end of the map, even if we have already decoded all
of the entries that were requested by the caller.

Signed-off-by: Sam Friedman <sam@golioth.io>