examples: Move fw_update reference implementation to samples
The FW Update implementation previously included in the core SDK
source is inteneded as a reference implementation. Users have
unique and diverse requirements for their firmware update logic,
and a one-size-fits-all solution is not feasible. To better reflect
the reference nature of the in-tree firmware update implementation,
we move the implementation out of the core SDK and into the
examples directory. As we move forward, this will allow us to
provide example implementations for firmware update that are
better matched to our supported platforms.

Signed-off-by: Sam Friedman <sam@golioth.io>