Add certificate rotation example
Adds an example demonstrating how to perform certificate rotation with the
PKI service. The example borrows heavily from the certificate provisioning
example, copying the code for lfs setup, the kconfig and devicetree
configuration, as well as test configuration, with the expectation that
we'll be changing this in the near future.

Signed-off-by: Trond Snekvik <trond@golioth.io>