client: zephyr: clean up credentials when destroying client
The Golioth client could not be recreated with new credentials after an
initial initialization with PKI, as the credentials slots were already in
use by the previous set of credentials. This would fail silently, and the
application would continue using the old certificate and private key
without their knowledge.

To fix this, and allow the golioth client to be destroyed, then recreated
with a different set of PKI credentials without rebooting, we have to
rearrange the init function to stop memsetting the client after the
configuration has been stored, so that we can determine which type of
credentials were used when we destroy the client. In the client destroy
call, we now additionally clear all credential slots occupied by the client
we're about to free, so that they're ready for use by a new client later.

The old client initialization function would wipe the configuration data
after the credentials were passed on to the TLS credential storage, but
it's unclear whether this was done intentionally as a way to avoid keeping
pointers to sensitive data around. The actual sensitive data was being
stored in separate buffers, and the wiped data only contained pointers to
said data. Clearing these pointers didn't actually do anything to improve
security, so I removed the memset altogether here.

This issue was resolved for PSKs in #385, but as that change addressed a
separate issue related to initialization, I'm unsure whether it can be
reverted as part of this cleanup.

Signed-off-by: Trond Snekvik <trond@golioth.io>