log: send log messages through pipelines
Updates the Golioth logging configuration to allow for sending logs to
Pipelines to be processed and delivered to the Golioth logs service or any
external destinations. The existing behavior of sending logs directly to
the Golioth logs service can be preserved by setting
CONFIG_GOLIOTH_LOG_LEGACY. If not using legacy logs ingestion, the
Pipelines path can be configured with CONFIG_GOLIOTH_LOG_PIPELINES_PATH,
but defaults to `logs`.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>